+++ src/apis/account.ts
@@ -1,53 +1,6 @@
 import { LoginResponse, RequestSignUpForm, UserAttendance, UserProfile } from '@@types/request';
-import axios from 'axios';
+import axios, { AxiosInstance } from 'axios';
 import { url } from '.';
-
-export const getEmailSecret = async (accessToken: string, emailValue: string) => {
-  // const response = await axios.get(
-  //     `${url}/cau_email`,
-  //     {
-  //         params: {
-  //             user_email: emailValue
-  //         },
-  //         headers: {
-  //             Authorization: `Bearer ${accessToken}`
-  //         }
-  //     }
-  // );
-  return { data: '1234' };
-};
-
-export const postEmailSecret = (accessToken: string, secretValue: string) => {
-  // const response = axios.post(
-  //     `${url}/cau_email`,
-  //     {
-  //         secret: secretValue
-  //     },
-  //     {
-  //         headers: {
-  //             Authorization: `Bearer ${accessToken}`
-  //         }
-  //     }
-  // ).then((res) => '1234');
-  return '1235';
-};
-
-export function postSignUpForm(form: RequestSignUpForm) {
-  return axios.post(
-    `${url}/signup`,
-    {
-      name: form.name,
-      generation: form.generation,
-      track: form.track,
-      is_admin: form.isAdmin,
-    },
-    {
-      headers: {
-        Authorization: `Bearer ${form.accessToken}`,
-      },
-    },
-  );
-}
 
 export function patchUserProfile(userProfile: UserProfile, accessToken: string) {
   return axios.patch(
@@ -116,7 +69,6 @@ export function login(code: string | string[]) {
 }
 
 export function getAcessToken(refresh_code: string) {
-  console.log('코드:', refresh_code);
   return axios
     .post(`/api/auths/token/refresh/`, {
       refresh: refresh_code,
